Revert "1.1.1 upstream release"
Stick with tracker 1.0 for now. First of all, the 1.1.1 release is
highly broken because it bumps the soname even though the 1.1/1.2
releases are supposed to be ABI compatible with 1.0. Secondly, the
tracker team is not sure at this point if GNOME 3.14 is going to go with
tracker 1.o or 1.2; so it's safer to stick with 1.0 for now.

This reverts commit 4641eb1d5653868725f8cbbf7cdabb8ffe8733df.